merge

+++ src/main/java/com/C_platform/item/repository/ImagesRepository.java
@@ -1,9 +1,10 @@
 package com.C_platform.item.repository;
 
+
 import com.C_platform.item.entity.Images;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ImagesRepository extends JpaRepository<Images, Integer> {
+public interface ImagesRepository extends JpaRepository<Images,Long> {
 }
