+++ src/main/java/com/C_platform/item/repository/ItemRepository.java
@@ -1,9 +1,10 @@
 package com.C_platform.item.repository;
 
+
 import com.C_platform.item.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item,Long> {
 }
